(PC-33780) feat(v2): remove TabBarInnerComponent v1
